# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Research Navigator API - Comprehensive scientific intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trends/{runId}:
    get:
      operationId: getTrends
      tags: [trends]
      summary: Get trend analysis for a collection run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trend analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendAnalysis"
  /trends/{runId}/compute:
    post:
      operationId: computeTrends
      tags: [trends]
      summary: Compute trend analysis for a collection run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trend analysis computed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendAnalysis"
...
components:
  schemas:
...
    Author:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        affiliations:
          type: array
          items:
            type: string
        citationCount:
          type: integer
        paperCount:
          type: integer
        hIndex:
          type: integer
      required:
        - id
        - name
        - affiliations
        - citationCount
        - paperCount
        - hIndex
...
    KeywordTrend:
      type: object
      properties:
        keyword:
          type: string
        dataPoints:
          type: array
          items:
            type: object
            properties:
              year:
                type: integer
              count:
                type: integer
              tfidfScore:
                type: number
            required:
              - year
              - count
              - tfidfScore
        growthRate:
          type: number
        peakYear:
          type: integer
      required:
        - keyword
        - dataPoints
        - growthRate
        - peakYear
    TopAuthor:
      type: object
      properties:
        author:
          $ref: "#/components/schemas/Author"
        paperCountInTopic:
          type: integer
        totalCitations:
          type: integer
      required:
        - author
        - paperCountInTopic
        - totalCitations
    PaperCluster:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        keywords:
          type: array
          items:
            type: string
        paperCount:
          type: integer
        growthRate:
          type: number
        description:
          type: string
      required:
        - id
        - label
        - keywords
        - paperCount
        - growthRate
        - description
    TrendAnalysis:
      type: object
      properties:
        runId:
          type: string
        topic:
          type: string
        keywordTrends:
          type: array
          items:
            $ref: "#/components/schemas/KeywordTrend"
        topAuthors:
          type: array
          items:
            $ref: "#/components/schemas/TopAuthor"
        clusters:
          type: array
          items:
            $ref: "#/components/schemas/PaperCluster"
        narrativeSummary:
          type: string
        totalPapersAnalyzed:
          type: integer
      required:
        - runId
        - topic
        - keywordTrends
        - topAuthors
        - clusters
        - narrativeSummary
        - totalPapersAnalyzed
...